Moved Spiral to Layouter folder
