+++ TagsCloudContainer/Layouter/CircularCloudLayouter.cs
@@ -6,14 +6,14 @@ namespace TagsCloudContainer.Layouter
 {
     public class CircularCloudLayouter
     {
-        private Spiral.Spiral _spiral;
+        private Spiral _spiral;
         private List<Rectangle> _rectangles;
         private Point _center;
 
         public CircularCloudLayouter(Point center)
         {
             _center = center;
-            _spiral = new Spiral.Spiral(center, 2);
+            _spiral = new Spiral(center, 2);
             _rectangles = new List<Rectangle>();
         }
